elrs/requirements.txt
pigpio==1.78
pyserial>=3.5
